requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openai>=1.6.0
pyyaml>=6.0
websockets>=12.0
httpx>=0.25.0
